Split build workflows for different platforms
Using one workflow to build for multiple platforms can lead
to timeout errors if one of them takes too long (especially when
there's emulation), and no images would be published even when
only one of them fails.

This split the build workflows into different files based on the
platform that publish the images to platform-specific nightly
channels, and add a new workflow to merge the platform-specific
nightlies into one nightly channel, so that
when one platform fails to build, others can still be updated.
This also allows us to build the arm64 images without emulation
which would be less likely to time out.