Fix permission and make ninja build cache work
This patch fixes the permission, so that the node distribution
from the local build as well as node-core-utils are installed
as the developer, not as root, which is less likely to incur
permission issues.

This also uses git-mtime-restore to match the mtime of the
project files with the timestamp from git, so that when a
devcontainer is mounted on the working directory, ninja
can easily pick up the cache already in the container and
enable fast incremental builds. Update the README accordingly
to provide information about how to mount the project to pick
up the build cache.